13 - django_by_example/mysite/requirements.txt
autopep8==1.4.4
Django==2.2.5
django-taggit==1.1.0
Markdown==3.1.1
psycopg2==2.8.3
pycodestyle==2.5.0
pytz==2019.2
sqlparse==0.3.0
